requirements.txt
fastapi==0.110.2
uvicorn==0.30.1

python-dotenv==0.21.1
python-json-logger==2.0.7
python-logstash==0.4.8